Renames yaml_util to swri_yaml_util. Refs #231.
