+++ swri_yaml_util/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(yaml_util)
+project(swri_yaml_util)
 
 find_package(catkin REQUIRED)
