+++ swri_yaml_util/package.xml
@@ -1,9 +1,9 @@
 <package format="2">
-  <name>yaml_util</name>
+  <name>swri_yaml_util</name>
   <version>0.0.1</version>
   <description>
 
-     yaml_util
+     swri_yaml_util
 
   </description>
   <author>Marc Alban</author>
